Move SSE function call example to correct examples directory
